+++ easybpms/src/main/java/com/easybpms/event/TaskExecutedObserver.java
@@ -54,7 +54,7 @@ public class TaskExecutedObserver implements Observer{
 		String inputParamValue = null;
 		try {
 			Method m = arg.getClass().getMethod("getId");
-			Integer i = (Integer)m.invoke(arg);
+			Long i = (Long)m.invoke(arg);
 			inputParamValue = i.toString(); 
 		} catch (Exception e) {
 			e.printStackTrace();
